create a branch for_porting_libcore

+++ Sources/Elastos/LibCore/src/elastos/utility/regex/CSplitter.cpp
@@ -158,7 +158,7 @@ ECode CSplitter::Split(
     Boolean result;
     String subStr;
 
-    while ((list.GetSize() + 1) != limit && (matcher->Find(&result), result)) {
+    while (((Int32)list.GetSize() + 1) != limit && (matcher->Find(&result), result)) {
         matcher->Start(&end);
         subStr = input.Substring(begin, end);
         list.PushBack(subStr);
